requirements.txt
beautifulsoup4==4.5.1
lxml==4.4.2
python-telegram-bot==12.7
requests==2.7.0
virtualenv==20.0.20